requirements.txt
# requirements.txt actualizado para la app completa
streamlit==1.42.2
pandas==2.2.3
numpy==2.2.3
plotly==6.0.0
pydeck==0.9.1
openpyxl==3.1.5
matplotlib==3.10.0
seaborn==0.13.2
scipy==1.13.1
Pillow==11.1.0
